backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.31.0
sympy>=1.13.0
networkx>=3.3
pydantic>=2.9.0
python-multipart>=0.0.12
pillow>=10.4.0
numpy>=1.26.0
google-generativeai>=0.8.0
